refactor(charts): Move data preparation for chart
